Support for ring 1.6+ asynchronous mode
This patch also changes the behavior of the handle-error parameter in
wrap-format-params.  Previously, the handler established an exception
handler around both the request body parsing actual handler evaluation.
The error handler is meant to be used as a hook for when the body
parsing fails, though, and thus it is now no longer called when the
actual handler throws an exception.